Update index.js
We want to update our code to include a section that includes a video which will include Elite Skills Trainer explaining the AI program. Also we need to update the instructions for how to use the training program. Place a " Begin Today" button under the How it works section right after the 5 steps. Lets move the articles section to the bottom of the page and change it to More Ball Handling Videos (which will be 3 youtube videos similar to the testimmonial section. Lastly, lets add a menu button at the top for the tracker that says "Upload" Update HERO section to turn Next-level results into Real Results